+++ src/server.rs
@@ -104,7 +104,7 @@ impl LanguageServer for Backend {
                     },
                 }),
                 completion_provider: Some(CompletionOptions {
-                    resolve_provider: Some(false),
+                    resolve_provider: Some(true),
                     trigger_characters: Some(vec![
                         "$".to_string(),
                         ">".to_string(),
@@ -479,6 +479,10 @@ impl LanguageServer for Backend {
 
     async fn completion(&self, params: CompletionParams) -> Result<Option<CompletionResponse>> {
         self.handle_completion(params).await
+    }
+
+    async fn completion_resolve(&self, params: CompletionItem) -> Result<CompletionItem> {
+        Ok(self.handle_completion_resolve(params))
     }
 
     async fn references(&self, params: ReferenceParams) -> Result<Option<Vec<Location>>> {